Crate frame_metadata

Decodable variant of the RuntimeMetadata.

This really doesn’t belong here, but is necessary for the moment. In the future it should be removed entirely to an external module for shimming on to the codec-encoded metadata.

Structs

DefaultByteGetter

Wrapper over dyn pointer for accessing a cached once byte value.

ErrorMetadata

All the metadata about a module error.

EventMetadata

All the metadata about an event.

ExtrinsicMetadata

Metadata of the extrinsic used by the runtime.

FnEncode

Newtype wrapper for support encoding functions (actual the result of the function).

FunctionArgumentMetadata

All the metadata about a function argument.

FunctionMetadata

All the metadata about a function.

ModuleConstantMetadata

All the metadata about one module constant.

ModuleMetadata

All metadata about an runtime module.

OuterEventMetadata

All the metadata about an outer event.

RuntimeMetadataPrefixed

Metadata prefixed by a u32 for reserved usage

RuntimeMetadataV13

The metadata of a runtime.

StorageEntryMetadata

All the metadata about one storage entry.

StorageMetadata

All metadata of the storage.

Enums

DecodeDifferent

A type that decodes to a different type than it encodes. The user needs to make sure that both types use the same encoding.

RuntimeMetadata

The metadata of a runtime. The version ID encoded/decoded through the enum nature of RuntimeMetadata.

RuntimeMetadataDeprecated

Enum that should fail.

StorageEntryModifier

A storage entry modifier.

StorageEntryType

A storage entry type.

StorageHasher

Hasher used by storage maps

Constants

META_RESERVED

Current prefix of metadata

Traits

DefaultByte

A technical trait to store lazy initiated vec value as static dyn pointer.

ModuleErrorMetadata

All the metadata about errors in a module.

Type Definitions

ByteGetter

Decode different for static lazy initiated byte value.

DecodeDifferentArray

RuntimeMetadataLastVersion

The latest version of the metadata.