frame_metadata

Module v16

Metadata v16

Re-exports

pub use super::v14::StorageEntryModifier;

pub use super::v14::StorageEntryType;

pub use super::v14::StorageHasher;

pub use super::v15::CustomMetadata;

pub use super::v15::CustomValueMetadata;

pub use super::v15::OuterEnums;

Structs

EnumDeprecationInfo

Deprecation information for enums in which specific variants can be deprecated. If the map is empty, then nothing is deprecated.

ExtrinsicMetadata

Metadata of the extrinsic used by the runtime.

PalletAssociatedTypeMetadata

Metadata of a pallet’s associated type.

PalletCallMetadata

Metadata for all calls in a pallet.

PalletConstantMetadata

Metadata about one pallet constant.

PalletErrorMetadata

Metadata about a pallet error.

PalletEventMetadata

Metadata about the pallet Event type.

PalletMetadata

All metadata about an runtime pallet.

PalletStorageMetadata

All metadata of the pallet’s storage.

PalletViewFunctionMetadata

Metadata about a pallet view function.

RuntimeApiMetadata

Metadata of a runtime trait.

RuntimeApiMethodMetadata

Metadata of a runtime method.

RuntimeMetadataV16

The metadata of a runtime.

StorageEntryMetadata

Metadata about one storage entry.

TransactionExtensionMetadata

Metadata of an extrinsic’s transaction extension.

Enums

ItemDeprecationInfo

Deprecation information for generic items.

VariantDeprecationInfo

Deprecation information for an item or variant in the metadata.

Type Aliases

FunctionParamMetadata

The metadata for a method or function parameter. This is identical to crate::v15::RuntimeApiMethodParamMetadata.

RuntimeMetadataLastVersion

Latest runtime metadata.