frame_metadata::v16

Type Alias RuntimeMetadataLastVersion

pub type RuntimeMetadataLastVersion = RuntimeMetadataV16;

Latest runtime metadata.

Aliased Type

pub struct RuntimeMetadataLastVersion {
    pub types: PortableRegistry,
    pub pallets: Vec<PalletMetadata<PortableForm>>,
    pub extrinsic: ExtrinsicMetadata<PortableForm>,
    pub apis: Vec<RuntimeApiMetadata<PortableForm>>,
    pub outer_enums: OuterEnums<PortableForm>,
    pub custom: CustomMetadata<PortableForm>,
}

Fields
types: PortableRegistry

Type registry containing all types used in the metadata.

pallets: Vec<PalletMetadata<PortableForm>>

Metadata of all the pallets.

extrinsic: ExtrinsicMetadata<PortableForm>

Metadata of the extrinsic.

apis: Vec<RuntimeApiMetadata<PortableForm>>

Metadata of the Runtime API.

outer_enums: OuterEnums<PortableForm>

The outer enums types as found in the runtime.

custom: CustomMetadata<PortableForm>

Allows users to add custom types to the metadata.