frame_metadata

Module v14

Metadata v14

Structs

ExtrinsicMetadata

Metadata of the extrinsic used by the runtime.

PalletCallMetadata

Metadata for all calls in a pallet

PalletConstantMetadata

Metadata about one pallet constant.

PalletErrorMetadata

Metadata about a pallet error.

PalletEventMetadata

Metadata about the pallet Event type.

PalletMetadata

All metadata about an runtime pallet.

PalletStorageMetadata

All metadata of the pallet’s storage.

RuntimeMetadataV14

The metadata of a runtime.

SignedExtensionMetadata

Metadata of an extrinsic’s signed extension.

StorageEntryMetadata

Metadata about one storage entry.

Enums

StorageEntryModifier

A storage entry modifier indicates how a storage entry is returned when fetched and what the value will be if the key is not present. Specifically this refers to the “return type” when fetching a storage entry, and what the value will be if the key is not present.

StorageEntryType

A type of storage value.

StorageHasher

Hasher used by storage maps

Constants

META_RESERVED

Current prefix of metadata

Type Aliases

RuntimeMetadataLastVersion

Latest runtime metadata